moved spade plot to spade.py
